moving code that should not be under /app to /test and small refactorings to clarify code and reduce heavy logs
