+++ kifi-backend/modules/common/app/com/keepit/scraper/ScraperServiceClient.scala
@@ -3,25 +3,25 @@ package com.keepit.scraper
 //import play.api.libs.concurrent.Execution.Implicits.defaultContext
 
 import java.io.IOException
-import com.keepit.model._
-import com.keepit.common.db.Id
-import com.keepit.common.service.{ RequestConsolidator, ServiceClient, ServiceType }
-import com.keepit.common.logging.Logging
-import com.keepit.common.healthcheck.AirbrakeNotifier
-import com.keepit.common.net.{ CallTimeouts, HttpClient }
-import com.keepit.common.zookeeper.ServiceCluster
+
 import com.google.inject.Inject
-import com.google.inject.util.Providers
-import com.keepit.common.routes.{ Common, Scraper }
+import com.keepit.common.amazon.AmazonInstanceInfo
+import com.keepit.common.db.Id
+import com.keepit.common.healthcheck.AirbrakeNotifier
+import com.keepit.common.logging.Logging
+import com.keepit.common.net.{ CallTimeouts, HttpClient }
+import com.keepit.common.routes.Scraper
+import com.keepit.common.service.{ RequestConsolidator, ServiceClient, ServiceType }
+import com.keepit.common.store.ImageSize
+import com.keepit.common.zookeeper.ServiceCluster
+import com.keepit.model._
+import com.keepit.scraper.embedly.EmbedlyInfo
+import com.keepit.scraper.extractor.ExtractorProviderType
 import com.keepit.search.Article
+import org.joda.time.DateTime
 import play.api.libs.functional.syntax._
 import play.api.libs.json._
-import com.keepit.scraper.extractor.ExtractorProviderType
-import com.keepit.common.amazon.AmazonInstanceInfo
-import org.joda.time.DateTime
-import akka.actor.Scheduler
-import com.keepit.scraper.embedly.EmbedlyInfo
-import com.keepit.common.store.ImageSize
+
 import scala.concurrent.Future
 import scala.concurrent.duration._
 
@@ -34,7 +34,7 @@ object ScrapeTuple {
 }
 
 case class ScrapeRequest(uri: NormalizedURI, scrapeInfo: ScrapeInfo, pageInfoOpt: Option[PageInfo], proxyOpt: Option[HttpProxy]) {
-  override def toString = s"(${uri.toShortString},${scrapeInfo.toShortString},${pageInfoOpt},$proxyOpt)"
+  override def toString = s"(${uri.toShortString},${scrapeInfo.toShortString},$pageInfoOpt,$proxyOpt)"
   def toShortString = s"(${uri.id},${scrapeInfo.id},${pageInfoOpt.flatMap(_.id)},${uri.url.take(50)}"
 }
 
@@ -103,7 +103,7 @@ case class ScraperThreadInstanceInfo(info: AmazonInstanceInfo, jobInfo: Either[S
   lazy val forkJoinThreadDetails: Seq[ScraperThreadDetails] = jobInfo match {
     case Left(threadDetails) => threadDetails
     case Right(details) =>
-      (details.lines filter { !_.isEmpty } toList) map { ScraperThreadDetails.buildFromString(_) } collect { case Some(x) => x }
+      (details.lines.toSeq.filterNot(_.isEmpty) map ScraperThreadDetails.buildFromString).flatten
   }
   lazy val scrapeThreadDetails: Seq[ScraperThreadDetails] = threadDetailsForTaskType(ScraperTaskType.SCRAPE)
   lazy val fetchBasicThreadDetails: Seq[ScraperThreadDetails] = threadDetailsForTaskType(ScraperTaskType.FETCH_BASIC)
@@ -254,35 +254,4 @@ class ScraperServiceClientImpl @Inject() (
       }
     }
   }
-}
-
-class FakeScraperServiceClientImpl(val airbrakeNotifier: AirbrakeNotifier, scheduler: Scheduler) extends ScraperServiceClient {
-
-  val serviceCluster: ServiceCluster = new ServiceCluster(ServiceType.TEST_MODE, Providers.of(airbrakeNotifier), scheduler, () => {})
-
-  protected def httpClient: com.keepit.common.net.HttpClient = ???
-
-  override def status(): Seq[Future[(AmazonInstanceInfo, Seq[ScrapeJobStatus])]] = Seq.empty
-
-  def getBasicArticle(url: String, proxy: Option[HttpProxy], extractor: Option[ExtractorProviderType]): Future[Option[BasicArticle]] = ???
-
-  def getSignature(url: String, proxy: Option[HttpProxy], extractor: Option[ExtractorProviderType]): Future[Option[Signature]] = ???
-
-  def getThreadDetails(filterState: Option[String]): Seq[Future[ScraperThreadInstanceInfo]] = ???
-
-  def getPornDetectorModel(): Future[Map[String, Float]] = ???
-
-  def detectPorn(query: String): Future[Map[String, Float]] = ???
-
-  def whitelist(words: String): Future[String] = ???
-
-  def getEmbedlyImageInfos(uriId: Id[NormalizedURI], url: String): Future[Seq[ImageInfo]] = ???
-
-  def getEmbedlyInfo(url: String): Future[Option[EmbedlyInfo]] = ???
-
-  def getURISummaryFromEmbedly(uri: NormalizedURI, minSize: ImageSize, descriptionOnly: Boolean): Future[Option[URISummary]] = ???
-
-  def getURIWordCount(uriId: Id[NormalizedURI], url: Option[String]): Future[Int] = ???
-
-  def getURIWordCountOpt(uriId: Id[NormalizedURI], url: Option[String]): Option[Int] = ???
 }
